requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
jiwer>=3.0.0
accelerate>=0.24.0
evaluate>=0.4.0
tensorboard>=2.14.0
scipy>=1.11.0
